22th commit (prepare to 2nd stage of Clean Architecture, some changes in SearchActivity)
