--- app/src/main/java/com/aleksagn/playlistmaker/ui/search/OnTrackClickListener.kt
@@ -1,8 +1,0 @@
-package com.aleksagn.playlistmaker.ui.search
-
-import android.content.Context
-import com.aleksagn.playlistmaker.domain.models.Track
-
-interface OnTrackClickListener {
-    fun onTrackClick(track: Track, context: Context)
-}
